Nova função para verificar se a música esta sendo usado em algum evento, antes de exclui-la. Foi usado o TDD para essa função

+++ AgendaVoluntarios/Controllers/MusicController.cs
@@ -145,6 +145,10 @@ namespace AgendaVoluntarios.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(Guid id)
         {
+            if (await _musicService.IsMusicLinkedToEvent(id))
+            {
+                return NotFound();
+            }
             await _musicService.DeleteAsync(id);
             return RedirectToAction(nameof(Index));
         }
